build abstraction layer around forking
`forkBackend()` creates a backend fork now.
