+++ main/background.js
@@ -1,56 +1,10 @@
-import { fork } from 'child_process';
-import createDebug from 'debug';
-import * as projectStore from '../lib/project-store';
-import * as userStore from '../lib/user-store';
 import dotenv from 'dotenv';
-
 dotenv.config();
 
 if (process.argv[2] === '--backend') {
-  createDebug.enable('backend*');
-  const debug = createDebug('backend');
-  debug('Backend worker initializing');
-  const { handleRequest } = require('../backend/handlers');
-
-  const processMessages = () =>
-    process.on('message', async ({ id, event, data }) => {
-      process.send({
-        id,
-        response: await handleRequest(event, data),
-      });
-    });
-
-  const handleInit = ({ event, dataDir }) => {
-    if (event === 'init') {
-      process.removeListener('message', handleInit);
-
-      projectStore.init(dataDir);
-      userStore.init(dataDir);
-
-      processMessages();
-      process.send('ready');
-    } else {
-      console.error('Invalid message received, expected `init` event.');
-    }
-  };
-
-  process.on('message', handleInit);
-  process.on('exit', () => debug('Background worker shutting down'));
+  const { initBackend } = require('./backend');
+  initBackend();
 } else {
-  const { app } = require('electron');
-  const { startApp } = require('./app');
-
-  const handleReady = message => {
-    if (message === 'ready') {
-      workerProcess.removeListener('message', handleReady);
-      startApp(workerProcess);
-    }
-  };
-
-  const workerProcess = fork(__filename, ['--backend']);
-  workerProcess.on('message', handleReady);
-  workerProcess.send({
-    event: 'init',
-    dataDir: app.getPath('userData'),
-  });
+  const { initApp } = require('./app');
+  initApp();
 }
